[bos #38643][EDF] Laplacian Smoothing 2D.
Add support to periodic faces to the laplacian smoother.

.

Add support to periodic faces to the centroidalSmooth.

Do not move singular nodes that are placed in the surface instead of on the edge as in the BodyFitting algorithm.

Small modif in averageByElement to not initialize theUVMap unproperly when theSurface.IsNull().